Fix stores and api doc
git-svn-id: https://subversion.assembla.com/svn/kinvey/ios-library/trunk@2181 868178e2-d38b-4ba8-a603-825f91c87007

Former-commit-id: f1054729935342a204582d12156e19cc9ac890e3

+++ KinveyKit/KinveyKit/Source/KCSAsyncConnection.h
@@ -35,7 +35,7 @@ typedef void (^RunBlock_t)();
 
 @property (nonatomic, readonly) NSInteger contentLength;
 @property (nonatomic, readonly) double percentComplete;
-@property (nonatomic) NSUInteger percentNotificationThreshold;
+@property (nonatomic) double percentNotificationThreshold;
